move qInstallMessageHandler to coreMain.h
